Add : Pinata as IPFS provider

+++ migrations/000001_create_content_records_table.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS content_records (
     sha256_hash VARCHAR(66) PRIMARY KEY,
     creator_address VARCHAR(42) NOT NULL,
-    phash BIGINT NOT NULL,
+    phash NUMERIC(20, 0) NOT NULL,
     timestamp BIGINT NOT NULL,
     ipfs_cid TEXT NOT NULL,
     ai_tool TEXT NOT NULL
